版本V1.1.2

+++ Model/Order.cs
@@ -22,7 +22,7 @@ namespace Model
         //订单名称
         public string order_name { get; set; }
         //下单数量
-        public Int64 order_num { get; set; }
+        public double order_num { get; set; }
         //下单时间(登记时间)
         public DateTime order_time { get; set; }
         //需求出货时间
@@ -38,18 +38,18 @@ namespace Model
         //单位
         public string unit { get; set; }
         //开单数量
-        public Int64 open_num { get; set; }
+        public double open_num { get; set; }
         //退单数量
-        public Int64 tui_num { get; set; }
+        public double tui_num { get; set; }
         //剩余数量
-        public Int64 remain_num { get; set; }
+        public double remain_num { get; set; }
         //图纸
         public string order_picture { get; set; }
         //插入时间
         public DateTime insert_time { get; set; }
 
         //real_num(用于判断是否加入到历史订单中)
-        public Int64 real_num { get; set; }
+        public double real_num { get; set; }
         //订单状态
         public int order_status { get; set; }
         //订单是否已开票
